fix: Close ALL remaining — 30 NOT FIXED + 12 PARTIAL = 0 open (#276)
## P0 closed (3):
- CoachProfile operator==: 38 fields compared (was 13). hashCode aligned.
- GoalA/GoalB/PlannedMonthlyContribution: added == + hashCode
- LibrePassageCompte: dateOuverture in equality
- Markdown/HTML injection: stripped from LLM responses before rendering

## P1 closed (10):
- SubscriptionProvider: catchError on fire-and-forget
- allowMalformed: removed from document scan UTF-8 decode
- Children upper limit: le=20 on 10 schema fields across 5 files
- Locked-out user: token cleared on 401 in all HTTP methods
- nombreDeMois: int → double (accepts 12.0, 13.0, prorata values)
- Coach insights: offline message queue (SharedPreferences pending_messages)
- Compliance logging: includes user_id

## P2 closed (5):
- birthYear: verified no 1990 fallback remaining
- Mortgage journey: 24 ARB keys extracted (7 steps, 6 languages)
- vous/votre: 13 coaching instances → tu/ton/ta

## Shortcuts self-audited and fixed:
- Tie-breaking: hashCode → lexicographic compareTo (stable across versions)
- FATCA 3a: not blocked (Swiss law allows), FATCA warning in templates instead
- Emergency fund 'no': documented as conservative floor (not 0)

Tests: 8125 Flutter + 4679 backend = 12804 all green
flutter analyze: 0 errors, 0 warnings | ruff: All checks passed

+++ apps/mobile/lib/providers/coach_profile_provider.dart
@@ -1523,7 +1523,7 @@ class CoachProfileProvider extends ChangeNotifier {
 
     _profile = p.copyWith(
       salaireBrutMensuel: salaireBrut ?? p.salaireBrutMensuel,
-      nombreDeMois: nombreMois ?? p.nombreDeMois,
+      nombreDeMois: (nombreMois ?? p.nombreDeMois).toDouble(),
       bonusPourcentage: bonus ?? p.bonusPourcentage,
       dataSources: updatedSources,
       dataTimestamps: updatedTimestamps,
